FINISH REFACTORING
